+++ src/main/java/domain/movement/ChariotMovement.java
@@ -2,6 +2,9 @@ package domain.movement;
 
 import domain.board.BoardState;
 import domain.board.Position;
+import domain.movement.vo.Direction;
+import domain.movement.vo.Path;
+import domain.movement.vo.Paths;
 import java.util.ArrayList;
 import java.util.List;
