+++ src/main/java/domain/movement/GeneralMovement.java
@@ -2,6 +2,9 @@ package domain.movement;
 
 import domain.board.BoardState;
 import domain.board.Position;
+import domain.movement.vo.Direction;
+import domain.movement.vo.Path;
+import domain.movement.vo.Paths;
 import java.util.List;
 
 public final class GeneralMovement implements Movement {
