+++ src/main/java/domain/movement/HorseMovement.java
@@ -2,6 +2,11 @@ package domain.movement;
 
 import domain.board.BoardState;
 import domain.board.Position;
+import domain.movement.vo.Delta;
+import domain.movement.vo.Direction;
+import domain.movement.vo.Directions;
+import domain.movement.vo.Path;
+import domain.movement.vo.Paths;
 import java.util.List;
 
 public final class HorseMovement implements Movement {
